Use a lazylistscope. #158
