+++ app/src/main/java/com/jerboa/ui/components/person/PersonProfileActivity.kt
@@ -27,7 +27,7 @@ import com.jerboa.db.AccountViewModel
 import com.jerboa.isScrolledToEnd
 import com.jerboa.openLink
 import com.jerboa.scrollToTop
-import com.jerboa.ui.components.comment.CommentNode
+import com.jerboa.ui.components.comment.CommentNodes
 import com.jerboa.ui.components.comment.edit.CommentEditViewModel
 import com.jerboa.ui.components.comment.edit.commentEditClickWrapper
 import com.jerboa.ui.components.comment.reply.CommentReplyViewModel
@@ -455,101 +455,99 @@ fun UserTabs(
                             modifier = Modifier.fillMaxSize()
                             // .simpleVerticalScrollbar(listState)
                         ) {
-                            items(nodes) { node ->
-                                CommentNode(
-                                    node = node,
-                                    onUpvoteClick = { commentView ->
-                                        account?.also { acct ->
-                                            personProfileViewModel.likeComment(
-                                                commentView = commentView,
-                                                voteType = VoteType.Upvote,
-                                                account = acct,
-                                                ctx = ctx,
-                                            )
-                                        }
-                                    },
-                                    onDownvoteClick = { commentView ->
-                                        account?.also { acct ->
-                                            personProfileViewModel.likeComment(
-                                                commentView = commentView,
-                                                voteType = VoteType.Downvote,
-                                                account = acct,
-                                                ctx = ctx,
-                                            )
-                                        }
-                                    },
-                                    onReplyClick = { commentView ->
-                                        commentReplyClickWrapper(
-                                            commentReplyViewModel = commentReplyViewModel,
-                                            parentCommentView = commentView,
-                                            postId = commentView.post.id,
-                                            navController = navController,
-                                        )
-                                    },
-                                    onSaveClick = { commentView ->
-                                        account?.also { acct ->
-                                            personProfileViewModel.saveComment(
-                                                commentView = commentView,
-                                                account = acct,
-                                                ctx = ctx,
-                                            )
-                                        }
-                                    },
-                                    onPersonClick = { personId ->
-                                        personClickWrapper(
-                                            personProfileViewModel,
-                                            personId,
-                                            account,
-                                            navController,
-                                            ctx
-                                        )
-                                    },
-                                    onCommunityClick = { community ->
-                                        communityClickWrapper(
-                                            communityViewModel = communityViewModel,
-                                            communityId = community.id,
-                                            account = account,
-                                            navController = navController,
+                            CommentNodes(
+                                nodes = nodes,
+                                onUpvoteClick = { commentView ->
+                                    account?.also { acct ->
+                                        personProfileViewModel.likeComment(
+                                            commentView = commentView,
+                                            voteType = VoteType.Upvote,
+                                            account = acct,
                                             ctx = ctx,
                                         )
-                                    },
-                                    onPostClick = { postId ->
-                                        postClickWrapper(
-                                            postViewModel = postViewModel,
-                                            postId = postId,
-                                            account = account,
-                                            navController = navController,
+                                    }
+                                },
+                                onDownvoteClick = { commentView ->
+                                    account?.also { acct ->
+                                        personProfileViewModel.likeComment(
+                                            commentView = commentView,
+                                            voteType = VoteType.Downvote,
+                                            account = acct,
                                             ctx = ctx,
                                         )
-                                    },
-                                    onEditCommentClick = { commentView ->
-                                        commentEditClickWrapper(
-                                            commentEditViewModel,
-                                            commentView,
-                                            navController,
+                                    }
+                                },
+                                onReplyClick = { commentView ->
+                                    commentReplyClickWrapper(
+                                        commentReplyViewModel = commentReplyViewModel,
+                                        parentCommentView = commentView,
+                                        postId = commentView.post.id,
+                                        navController = navController,
+                                    )
+                                },
+                                onSaveClick = { commentView ->
+                                    account?.also { acct ->
+                                        personProfileViewModel.saveComment(
+                                            commentView = commentView,
+                                            account = acct,
+                                            ctx = ctx,
                                         )
-                                    },
-                                    onReportClick = { commentView ->
-                                        commentReportClickWrapper(
-                                            createReportViewModel,
-                                            commentView.comment.id,
-                                            navController,
+                                    }
+                                },
+                                onPersonClick = { personId ->
+                                    personClickWrapper(
+                                        personProfileViewModel,
+                                        personId,
+                                        account,
+                                        navController,
+                                        ctx
+                                    )
+                                },
+                                onCommunityClick = { community ->
+                                    communityClickWrapper(
+                                        communityViewModel = communityViewModel,
+                                        communityId = community.id,
+                                        account = account,
+                                        navController = navController,
+                                        ctx = ctx,
+                                    )
+                                },
+                                onPostClick = { postId ->
+                                    postClickWrapper(
+                                        postViewModel = postViewModel,
+                                        postId = postId,
+                                        account = account,
+                                        navController = navController,
+                                        ctx = ctx,
+                                    )
+                                },
+                                onEditCommentClick = { commentView ->
+                                    commentEditClickWrapper(
+                                        commentEditViewModel,
+                                        commentView,
+                                        navController,
+                                    )
+                                },
+                                onReportClick = { commentView ->
+                                    commentReportClickWrapper(
+                                        createReportViewModel,
+                                        commentView.comment.id,
+                                        navController,
+                                    )
+                                },
+                                onBlockCreatorClick = {
+                                    account?.also { acct ->
+                                        personProfileViewModel.blockPerson(
+                                            person = it,
+                                            account = acct,
+                                            ctx = ctx,
                                         )
-                                    },
-                                    onBlockCreatorClick = {
-                                        account?.also { acct ->
-                                            personProfileViewModel.blockPerson(
-                                                person = it,
-                                                account = acct,
-                                                ctx = ctx,
-                                            )
-                                        }
-                                    },
-                                    showPostAndCommunityContext = true,
-                                    account = account,
-                                    moderators = listOf()
-                                )
-                            }
+                                    }
+                                },
+                                showPostAndCommunityContext = true,
+                                account = account,
+                                moderators = listOf()
+                            )
                         }
                     }
                 }
